Initial World and Entity implementation
Initial World and Entity implementation.

The Entity class is a basic game object.

The CPP class is similar to the FP class in FlashPunk. It holds static
data related to the game.

The BaseWorld class is a "scene" that will be subclassed and implemented
with logic and drawing.

The Engine class is simply a wrapper for initializing and starting
openFrameworks and OpenGL.

The StateHandler class manages update loops for the particular active
World.